The main highlight of this change is that the test (direct connection) and JNDI data sources have been split out into their own XML Spring Bean files with corresponding properties files.
The main application Spring Beans XML file and properties file now only contain beans/properties which are not related to the data source.

As an additional clean-up some of the values of the beans were changed to property placeholders and their values placed in property files.

.classpath
- Added the other Jars to the classpath which are used in the web components.

*.launch
- corrected the location of the log4j configuration and removed the MODE env var

CommandLineMain
- added the test_datasource.xml to the configuration locations loaded

web.xml
- added the jndi_datasource.xml into the locations of the context configurations

*.properties, *.xml
- Fully split out the test and jndi datasources into separate XML files and also split out their properties.  Now the main
application Spring Beans are in footballmanagerroles.xml and properties in footballmanagerroles.propertie.  The jndi datasource
beans are in jndi_datasource.xml with properties in jndi_datasource.properties.  The test datasource (a direct connection) have
beans in test_datasource.xml and properties in test_datasource.properties

*Test.java
- added the test_datasource.xml into the list of locations for Spring Beans to be loaded from.