Changed SimpleHTMLFootballManagerRolesView to get the top/bottom of the HTML page from existing HTML files so that the overall appearance is more professional.  This meant the beans file was updated too.
Had to update the build.xml to copy the files from the main/war dir to the WEB-INF/classes dir during the build so that they might be accessed from the code.

RecommendRole has to reset the name of the player since it downloads it as a temporary file.  The temporary file name is now timestamped too.

Project is complete for all intents and purposes.